consul/proxy/requirements.txt
Flask>=1.1.1
structlog
pyyaml
readerwriterlock
attrs
